+++ cmd/linstor-affinity-controller/main.go
@@ -10,12 +10,15 @@ import (
 	k8scli "k8s.io/component-base/cli"
 
 	"github.com/piraeusdatastore/linstor-affinity-controller/pkg/controller"
+	"github.com/piraeusdatastore/linstor-affinity-controller/pkg/leaderelection"
 	"github.com/piraeusdatastore/linstor-affinity-controller/pkg/version"
 )
 
 func NewControllerCommand() *cobra.Command {
 	cfgflags := genericclioptions.NewConfigFlags(false)
 	var reconcileRate, resyncRate, timeout time.Duration
+	var electorCfg leaderelection.Config
+	var bindAddress string
 
 	cmd := &cobra.Command{
 		Use:     "linstor-volume-controller",
@@ -29,11 +32,18 @@ func NewControllerCommand() *cobra.Command {
 				return err
 			}
 
+			elector, err := electorCfg.MakeElector(cancel, cfg)
+			if err != nil {
+				return err
+			}
+
 			ctrl, err := controller.NewReconciler(&controller.Config{
 				RestCfg:       cfg,
 				ResyncRate:    resyncRate,
 				ReconcileRate: reconcileRate,
 				Timeout:       timeout,
+				LeaderElector: elector,
+				BindAddress:   bindAddress,
 			})
 			if err != nil {
 				return err
@@ -47,6 +57,9 @@ func NewControllerCommand() *cobra.Command {
 	cmd.Flags().DurationVar(&reconcileRate, "reconcile-rate", 15*time.Second, "how often the cluster state should be reconciled")
 	cmd.Flags().DurationVar(&resyncRate, "resync-rate", 5*time.Minute, "how often the internal object cache should be resynchronized")
 	cmd.Flags().DurationVar(&timeout, "timeout", 1*time.Minute, "how long a single reconcile attempt can take")
+	cmd.Flags().StringVar(&bindAddress, "bind-address", "[::]:8000", "the address to use for /healthz and /readyz probes")
+	electorCfg.AddFlags(cmd.Flags())
+
 	return cmd
 }